// types.ts
export enum Panel {
  FILES = 'FILES',
  SEARCH = 'SEARCH',
  CHAT = 'CHAT',
  SOURCE_CONTROL = 'SOURCE_CONTROL',
  SECURITY = 'SECURITY',
  TERRAFORM = 'TERRAFORM',
  DOCKER = 'DOCKER',
  EXTENSIONS = 'EXTENSIONS',
  TEAM = 'TEAM',
  INTEGRATIONS = 'INTEGRATIONS',
  AI_MARKETPLACE = 'AI_MARKETPLACE',
  PERSONAS = 'PERSONAS',
  STITCH_STUDIO = 'STITCH_STUDIO',
  RECORDER = 'RECORDER',
  SETTINGS = 'SETTINGS',
  HELP = 'HELP',
  DEBUG = 'DEBUG',
  TESTING = 'TESTING',
  DATABASE = 'DATABASE',
  KANBAN = 'KANBAN',
  PREVIEW = 'PREVIEW',
  API_STUDIO = 'API_STUDIO',
  WHITEBOARD = 'WHITEBOARD',
  DOCUMENT = 'DOCUMENT',
  TERMINAL = 'TERMINAL',
  DEPLOYMENT = 'DEPLOYMENT',
  ML_STUDIO = 'ML_STUDIO',
  PRODUCTION_ENGINEER = 'PRODUCTION_ENGINEER'
}